Mis eventos
Se visualiza correctamente los eventos creados y a los que fue invitado,
con sus respectivos nombre de organizador y opciones.

+++ app/views/eventos/MisEventos.blade.php
@@ -38,31 +38,66 @@
             		<th>ACCIONES</th>
             	</thead>
             	<tbody>
-            		@foreach ($listaDeEventos as $evento )
+            	@foreach ($listaDeEventos as $evento )
 					@if ($evento->creador == Session::get('usuario_id'))
-            		<tr>
+						<tr>
             				<td>{{$evento->nombre}}</td>
             				<td>{{$evento->direccion}}</td>
             				<td>{{$evento->fecha}}</td>
-            				<td>{{$evento->creador}}</td> 
+            				<td>Usted</td> 
             				<td>
-								<a class="btn btn-success" href="/verevento/{{$evento->id}}"  title="Vista del evento"><span class="glyphicon glyphicon-eye-open"></span>
-								<i class="icon-envelope"></i>Ver
-								</a>
+								@if ($evento->creador == Session::get('usuario_id'))
+										<a class="btn btn-success" href="/verevento/{{$evento->id}}"  title="Vista del evento"><span class="glyphicon glyphicon-eye-open"></span>
+										<i class="icon-envelope"></i>Ver
+										</a>
 								
-								<a class="btn btn-info" href="/modificarevento/{{$evento->id}}" title="Modifica el evento">
-								<i class="icon-envelope"></i>Modificar
-								</a>
+										<a class="btn btn-info" href="/modificarevento/{{$evento->id}}" title="Modifica el evento">
+										<i class="icon-envelope"></i>Modificar
+										</a>
 								
-								<a class="btn btn-danger" href="eliminarevento/{{$evento->id}}" title="Borra el evento">
-								<i class="icon-envelope"></i>Eliminar
-								</a>
-								
+										<a class="btn btn-danger" href="eliminarevento/{{$evento->id}}" title="Borra el evento">
+										<i class="icon-envelope"></i>Eliminar
+										</a>
+									
+								@endif								
 								</span>
 							</td>		
-            		</tr>
-					@endif 
-            		@endforeach
+						</tr>
+					@else
+						@foreach($listaDeInvitados as $invitado)
+							@if ($invitado->idusuario == Session::get('usuario_id'))
+									@if ($invitado->idevento == $evento->id)
+									<tr>
+										<td>{{$evento->nombre}}</td>
+										<td>{{$evento->direccion}}</td>
+										<td>{{$evento->fecha}}</td>
+										<td>
+										@foreach ($listaDeUsuarios as $usuario)
+											@if ($usuario->id==$evento->creador)
+												
+											{{$usuario->username}}
+											@endif
+										@endforeach
+										</td> 
+										<td>
+								
+										<a class="btn btn-success" href="/verevento/{{$evento->id}}"  title="Vista del evento"><span class="glyphicon glyphicon-eye-open"></span>
+										<i class="icon-envelope"></i>Ver
+										</a>
+																
+										</span>
+										</td>		
+									</tr>
+									@endif
+									
+							@endif
+						@endforeach
+											
+					@endif
+				@endforeach
+					
+					
+            		
             	</tbody>
             </table>
